Wrote the ConditionProcessorr but there's one nasty, thorny issue left:  I was hoping that I could transparently modify condition definitions better than I do for SpecTcl by doing a replace of the underlying RefCell in the dictionary from which all geting is done...unfortunately - I'm not able to figure out how to make the RefCell::replace call callable because I can't seem to figure out the trait bounds that it needs to properly operate.
Where I'm at now is that a condition replacement will invalidate
all references to that gate.  This is unsatisfying as it's worse that
SpecTcl's gate handling

I wonder if there's a solution by binding the gate dictionary into
the spectra and, when a gate becomes invalidated try to look it up
again -- this requires the same sort of thing to be done with the
compound conditions maintaining the names of those conditions as well
as the conditions themselves.