Issue #67
This seems complex since, in windows, the lock on the file
is retained while the map is live, and thread exit is asynchronous.

*   Provide shared memory method to return the path to the file.
*   Have the exit request retun that path.
*   In exit wait a bit (might be able to use a join handle here
actually) - then remove the file.

That appears to work. Note that in Linux it seems the file is a true
temp (delete on last close) file and does not get left behind.
This may lead to a message from linux rustogramer that the file
could not be deleted because it was no longer there.