Issue #63 - Extend unit tests to ReST requrests.
Comleted testing ASCII sread:
Filed an issue on the fact that at present, sread format = ascii has
no way to return error information tot he client.  It properly reads no
spectra but it really should return an error.
This is a separate issue because the program behavior is, nevertheless
reasonable so fixing this defect is a lower priority issue.