Issue #63 - Extend test coverage to ReST handlers.
*  Fix error in ASCII header processing.  in the case where there are
two sets of parameters:
("p1" "p2" ) ("p3" "p4" )

The two parentheses were split into components differently and the
second one was done incorrectlyh resulting in an extra 'parameter' named
with an empty string.  This resulted in a panic when attempting to
remove the quotes from around it.

* Tested particle gamma spectrum write/recovery.  This brought to light
the bug above.